+++ hdf/test/sdmms.c
@@ -94,11 +94,11 @@ test_sdmms(void)
         ui32scale[i] = (uint32)((i * 400000000U) + j);    /* range: 0 ~ 4-billion */
     }
 
-    cal1 = (float64)10.0;
-    cal2 = (float64)0.0;
-    cal3 = (float64)27.0;
-    cal4 = (float64)1.5;
-    cal5 = (int32)DFNT_INT16;
+    cal1 = 10.0;
+    cal2 = 0.0;
+    cal3 = 27.0;
+    cal4 = 1.5;
+    cal5 = DFNT_INT16;
 
     ret = DFSDsetdims(rank, dims);
     RESULT("DFSDsetdims");
