+++ hdf/test/tvsfpack.c
@@ -80,8 +80,8 @@ fpack(void)
     pntr = databuf;
     /* pack a record at a time */
     for (i = 0; i < NRECORDS; i++) {
-        source[i].temp   = (float32)1.11 * (float32)(i + 1);
-        source[i].height = (float32)2.22 * (float32)(i + 1);
+        source[i].temp   = 1.11F * (float32)(i + 1);
+        source[i].height = 2.22F * (float32)(i + 1);
         source[i].speed  = (int16)i;
         source[i].ident  = (char)('A' + i);
         /* test error checks  */
@@ -133,8 +133,8 @@ fpack(void)
     /* pack a field at a time */
     pntr = databuf;
     for (i = 0; i < NRECORDS; i++) {
-        source[i].temp   = (float32)3.33 * (float32)(i + 1);
-        source[i].height = (float32)4.44 * (float32)(i + 1);
+        source[i].temp   = 3.33F * (float32)(i + 1);
+        source[i].height = 4.44F * (float32)(i + 1);
         source[i].speed  = (int16)(2 * i);
         source[i].ident  = (char)('a' + i);
         databufptr[0]    = &source[i].speed;
@@ -223,9 +223,8 @@ funpack(void)
     }
 
     for (i = 0; i < NRECORDS; i++)
-        if ((iident[i] != (char)('A' + i)) ||
-            (fabs((double)(itemp[i] - (float32)1.11 * (float32)(i + 1))) > EPS) || (ispeed[i] != i) ||
-            (fabs((double)(iheight[i] - (float32)2.22 * (float32)(i + 1))) > EPS)) {
+        if ((iident[i] != (char)('A' + i)) || (fabs((double)(itemp[i] - 1.11F * (float32)(i + 1))) > EPS) ||
+            (ispeed[i] != i) || (fabs((double)(iheight[i] - 2.22F * (float32)(i + 1))) > EPS)) {
             num_errs++;
             printf(">>> Wrong data1 after VSfpack.\n");
         }
@@ -250,9 +249,8 @@ funpack(void)
     }
 
     for (i = 0; i < NRECORDS; i++)
-        if ((iident[i] != (char)('a' + i)) ||
-            (fabs((double)(itemp[i] - (float32)3.33 * (float32)(i + 1))) > EPS) || (ispeed[i] != 2 * i) ||
-            (fabs((double)(iheight[i] - (float32)4.44 * (float32)(i + 1))) > EPS)) {
+        if ((iident[i] != (char)('a' + i)) || (fabs((double)(itemp[i] - 3.33F * (float32)(i + 1))) > EPS) ||
+            (ispeed[i] != 2 * i) || (fabs((double)(iheight[i] - 4.44F * (float32)(i + 1))) > EPS)) {
             num_errs++;
             printf(">>> Wrong data2 after VSfpack.\n");
         }
@@ -285,7 +283,7 @@ funpack(void)
     }
     for (i = 0; i < NRECORDS; i++)
         if ((iident[i] != 'A' + i) || (ispeed[i] != i) ||
-            (fabs((double)(iheight[i] - (float32)2.22 * (float32)(i + 1))) > EPS)) {
+            (fabs((double)(iheight[i] - 2.22F * (float32)(i + 1))) > EPS)) {
             num_errs++;
             printf(">>> Wrong subset data1 after VSfpack.\n");
         }
@@ -303,7 +301,7 @@ funpack(void)
     }
     for (i = 0; i < NRECORDS; i++)
         if ((iident[i] != (char)('a' + i)) || (ispeed[i] != 2 * i) ||
-            (fabs((double)(iheight[i] - (float32)4.44 * (float32)(i + 1))) > EPS)) {
+            (fabs((double)(iheight[i] - 4.44F * (float32)(i + 1))) > EPS)) {
             num_errs++;
             printf(">>> Wrong subset data2 after VSfpack.\n");
         }
