+++ hdf/util/fp2hdf.c
@@ -1197,7 +1197,7 @@ indexes(float32 *scale, int dim, int *idx, int res)
         goto err;
     }
     for (i = 0; i < dim - 1; i++)
-        midpt[i] = (scale[i] + scale[i + 1]) * (float32)0.5;
+        midpt[i] = (scale[i] + scale[i + 1]) * 0.5F;
     midpt[dim - 1] = scale[dim - 1] + (scale[dim - 1] - midpt[dim - 2]);
 
     /*
@@ -1261,8 +1261,8 @@ interp(struct Input *in, struct Raster *im)
     float32        loc;
     float32        range;
     float32        ratio;
-    float32        hrange, vrange, drange = (float32)0.0;
-    float32        hdelta, vdelta, ddelta = (float32)0.0;
+    float32        hrange, vrange, drange = 0.0F;
+    float32        hdelta, vdelta, ddelta = 0.0F;
     float32        t1, t2, t3, t4, t5, t6;
     float32       *hratio, *vratio, *dratio = NULL;
     float32       *pt[8];
@@ -1274,7 +1274,7 @@ interp(struct Input *in, struct Raster *im)
      * determine the range of pixel locations
      */
     range  = in->max - in->min;
-    ratio  = (float32)237.9 / range;
+    ratio  = 237.9F / range;
     hrange = in->hscale[in->dims[0] - 1] - in->hscale[0];
     vrange = in->vscale[in->dims[1] - 1] - in->vscale[0];
     if (in->rank == 3)
@@ -1333,7 +1333,7 @@ interp(struct Input *in, struct Raster *im)
         loc     = hdelta * (float)i + in->hscale[0];
         hinc[i] = 0;
         while ((j < (in->dims[0] - 2)) &&
-               ((hrange > (float32)0.0) ? (in->hscale[j + 1] < loc) : (in->hscale[j + 1] > loc))) {
+               ((hrange > 0.0F) ? (in->hscale[j + 1] < loc) : (in->hscale[j + 1] > loc))) {
             hinc[i] += 1;
             j += 1;
         }
@@ -1342,7 +1342,7 @@ interp(struct Input *in, struct Raster *im)
     for (i = 0, j = 0, voff[0] = 0; i < im->vres; i++) {
         loc = vdelta * (float)i + in->vscale[0];
         while ((j < (in->dims[1] - 2)) &&
-               ((vrange > (float32)0.0) ? (in->vscale[j + 1] < loc) : (in->vscale[j + 1] > loc))) {
+               ((vrange > 0.0F) ? (in->vscale[j + 1] < loc) : (in->vscale[j + 1] > loc))) {
             voff[i] += 1;
             j += 1;
         }
@@ -1353,7 +1353,7 @@ interp(struct Input *in, struct Raster *im)
         for (i = 0, j = 0, doff[0] = 0; i < im->dres; i++) {
             loc = ddelta * (float)i + in->dscale[0];
             while ((j < (in->dims[2] - 2)) &&
-                   ((drange > (float32)0.0) ? (in->dscale[j + 1] < loc) : (in->dscale[j + 1] > loc))) {
+                   ((drange > 0.0F) ? (in->dscale[j + 1] < loc) : (in->dscale[j + 1] > loc))) {
                 doff[i] += 1;
                 j += 1;
             }
@@ -1383,7 +1383,7 @@ interp(struct Input *in, struct Raster *im)
                     pix = in->max; /* clip (bug fix) */
                 if (pix < in->min)
                     pix = in->min; /* ditto */
-                *ip++ = (unsigned char)((ratio * (pix - in->min)) + (float32)1.5);
+                *ip++ = (unsigned char)((ratio * (pix - in->min)) + 1.5F);
             }
         }
     }
@@ -1412,7 +1412,7 @@ interp(struct Input *in, struct Raster *im)
                         pix = in->max; /* clip (bug fix) */
                     if (pix < in->min)
                         pix = in->min; /* ditto */
-                    *ip++ = (unsigned char)((ratio * (pix - in->min)) + (float32)1.5);
+                    *ip++ = (unsigned char)((ratio * (pix - in->min)) + 1.5F);
                 }
             }
         }
@@ -1596,7 +1596,7 @@ pixrep(struct Input *in, struct Raster *im)
     dp    = (float32 *)in->data;
     ip    = im->image;
     range = in->max - in->min;
-    ratio = (float32)237.9 / range;
+    ratio = 237.9F / range;
 
     /*
      * determine the scale indexes of the horizontal pixel locations
@@ -1653,7 +1653,7 @@ pixrep(struct Input *in, struct Raster *im)
                  */
                 if (vidx[j] > ovidx) {
                     for (i = 0; i < in->dims[0]; i++)
-                        pix[i] = (unsigned char)((ratio * (*dp++ - in->min)) + (float32)1.5);
+                        pix[i] = (unsigned char)((ratio * (*dp++ - in->min)) + 1.5F);
                     for (i = 0; i < im->hres; i++)
                         *ip++ = pix[hidx[i]];
                     /*
